requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1.1
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
django-pgconnection==1.0.1
django-pgtrigger==2.1.0
django-reversion==3.0.8
djangorestframework==3.11.1
djangorestframework-simplejwt==4.3.0
psycopg2==2.8.6
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Unidecode==1.1.1
